requirements.txt
boto3==1.26.94
PyYAML==6.0